feat: 필요없는 파일 의존성 제거 및 파일 삭제
