--- Projects/Feature/Sent/Sources/CreateEnvelopeAdditional/CreateEnvelopeAdditionalContact/CreateEnvelopeAdditionalContactHelper.swift
@@ -1,14 +1,0 @@
-//
-//  CreateEnvelopeAdditionalContactHelper.swift
-//  Sent
-//
-//  Created by MaraMincho on 5/10/24.
-//  Copyright © 2024 com.oksusu. All rights reserved.
-//
-
-import Foundation
-
-struct CreateEnvelopeAdditionalContactHelper: Equatable {
-  var textFieldText: String = ""
-  var isHighlight: Bool = false
-}
