--- Projects/Feature/Sent/Sources/CreateEnvelopeAdditional/CreateEnvelopeAdditionalIsGift/CreateEnvelopeAdditionalIsGiftPropertyHelper.swift
@@ -1,14 +1,0 @@
-//
-//  CreateEnvelopeAdditionalIsGiftPropertyHelper.swift
-//  Sent
-//
-//  Created by MaraMincho on 5/10/24.
-//  Copyright © 2024 com.oksusu. All rights reserved.
-//
-
-import Foundation
-
-struct CreateEnvelopeAdditionalIsGiftPropertyHelper: Equatable {
-  var textFieldText: String = ""
-  init() {}
-}
