--- Projects/Feature/Sent/Sources/CreateEnvelopeAdditional/CreateEnvelopeAdditionalMemo/CreateEnvelopeAdditionalMemoHelper.swift
@@ -1,14 +1,0 @@
-//
-//  CreateEnvelopeAdditionalMemoHelper.swift
-//  Sent
-//
-//  Created by MaraMincho on 5/10/24.
-//  Copyright © 2024 com.oksusu. All rights reserved.
-//
-
-import Foundation
-
-struct CreateEnvelopeAdditionalMemoHelper: Equatable {
-  var textFieldText: String = ""
-  var isHighlight: Bool = false
-}
