+++ Projects/Feature/Sent/Sources/SentSerach/SentSearchNetwork.swift
@@ -15,12 +15,26 @@ import SSNetwork
 // MARK: - SentSearchNetwork
 
 struct SentSearchNetwork {
-  @Dependency(\.createEnvelopeNameNetwork) var searchFriendsNetwork
   @Dependency(\.sentMainNetwork) var searchAmountEnvelopeNetwork
   @Dependency(\.sentMainNetwork) var searchEnvelopeNetwork
 
+  private let provider = MoyaProvider<Network>(session: .init(interceptor: SSTokenInterceptor.shared))
+
   func searchFriendsBy(name: String) async throws -> [SentSearchItem] {
-    return try await searchFriendsNetwork.searchFriendBy(name: name)
+    let data: PageResponseDtoSearchFriendResponse = try await provider.request(.searchFriend(name: name))
+    return data.data.compactMap { dto -> SentSearchItem? in
+      guard let recentEnvelope = dto.recentEnvelope,
+            let targetDate = CustomDateFormatter.getYearAndMonthDateString(from: dto.recentEnvelope?.handedOverAt)
+      else {
+        return nil
+      }
+      return .init(
+        id: dto.friend.id,
+        title: dto.friend.name,
+        firstContentDescription: dto.recentEnvelope?.category,
+        secondContentDescription: targetDate
+      )
+    }
   }
 
   func searchEnvelopeBy(amount: Int64) async throws -> [SentSearchItem] {
@@ -36,11 +50,116 @@ struct SentSearchNetwork {
 
 extension SentSearchNetwork: DependencyKey {
   static var liveValue: SentSearchNetwork = .init()
+
+  enum Network: SSNetworkTargetType {
+    case searchFriend(name: String?)
+
+    var additionalHeader: [String: String]? { nil }
+    var path: String { "friends" }
+    var method: Moya.Method {
+      switch self {
+      case .searchFriend:
+        return .get
+      }
+    }
+
+    var task: Moya.Task {
+      switch self {
+      case let .searchFriend(name: name):
+        if let name {
+          return .requestParameters(parameters: ["name": name], encoding: URLEncoding.queryString)
+        }
+        return .requestPlain
+      }
+    }
+  }
 }
 
 extension DependencyValues {
   var sentSearchNetwork: SentSearchNetwork {
     get { self[SentSearchNetwork.self] }
     set { self[SentSearchNetwork.self] = newValue }
+  }
+}
+
+// MARK: - PageResponseDtoSearchFriendResponse
+
+struct PageResponseDtoSearchFriendResponse: Codable {
+  let data: [SearchFriendResponse]
+  let page: Int?
+  let size: Int?
+  let totalPage: Int
+  let totalCount: Int
+  let sort: SortObject
+
+  enum CodingKeys: CodingKey {
+    case data
+    case page
+    case size
+    case totalPage
+    case totalCount
+    case sort
+  }
+}
+
+// MARK: - SearchFriendResponse
+
+struct SearchFriendResponse: Codable {
+  let friend: FriendModel
+  let relationship: RelationshipInfoModel
+  let recentEnvelope: RecentEnvelopeModel?
+  enum CodingKeys: CodingKey {
+    case friend
+    case relationship
+    case recentEnvelope
+  }
+}
+
+// MARK: - FriendModel
+
+struct FriendModel: Codable {
+  let id: Int64
+  let name: String
+  let phoneNumber: String?
+  enum CodingKeys: CodingKey {
+    case id
+    case name
+    case phoneNumber
+  }
+}
+
+// MARK: - RecentEnvelopeModel
+
+struct RecentEnvelopeModel: Codable {
+  let category, handedOverAt: String
+  enum CodingKeys: CodingKey {
+    case category
+    case handedOverAt
+  }
+}
+
+// MARK: - RelationshipInfoModel
+
+struct RelationshipInfoModel: Codable {
+  let id: Int
+  let relation: String
+  let customRelation: String?
+  let description: String?
+  enum CodingKeys: CodingKey {
+    case id
+    case relation
+    case customRelation
+    case description
+  }
+}
+
+// MARK: - SortObject
+
+struct SortObject: Codable {
+  let empty, unsorted, sorted: Bool
+  enum CodingKeys: CodingKey {
+    case empty
+    case unsorted
+    case sorted
   }
 }
